Ensure default customization of Chart dependencies (#748)
### Description of the change

This adds unittests for the default configuration of our dependencies.

### Benefits

We are able to detect changes in upstream Charts for those settings.

### Applicable issues

- fixes #409

### Additional information

I merged existing unittests with the new ones for `postgresql-ha` and `postgresql`.
However, I decided to not do this for `redis-cluster` and `redis`, because the existing tests[^1] also ensure how the Chart behaves when both redis dependencies are disabled. It didn't felt right to split those files.

### Checklist

- [x] Templating unittests are added

Signed-off-by: justusbunsi <sk.bunsenbrenner@gmail.com>

[^1]: https://gitea.com/gitea/helm-chart/src/commit/e733287dc236e34a6caebf6bd8de5e6f25533eae/unittests/config/cache-config.yaml, https://gitea.com/gitea/helm-chart/src/commit/e733287dc236e34a6caebf6bd8de5e6f25533eae/unittests/config/queue-config.yaml, https://gitea.com/gitea/helm-chart/src/commit/e733287dc236e34a6caebf6bd8de5e6f25533eae/unittests/config/session-config.yaml

Reviewed-on: https://gitea.com/gitea/helm-chart/pulls/748
Co-authored-by: justusbunsi <sk.bunsenbrenner@gmail.com>
Co-committed-by: justusbunsi <sk.bunsenbrenner@gmail.com>